refactor(Backend): separated logic from the controller
